Edit sim to convert raw images for convenience.
Still haven't found an nn conf but just letting that fail for now while
I try to get the camera pipeline to look right.

Currently, it seems to be turning a (480, 640, 3) PNG image of a tree
into a speckled mostly-white image with some recognizable features from
the original image, just with the wrong colors everywhere. Not sure if
this is a bug in the camera pipeline implementation or a mistake in,
say, the size or format of the inputs I'm providing.