nnet: Fix a memory bug in decompression and quantization.
When the weights are compressed, we use the decompression block to
decompress the weights before invoking the inner product block.
Since the decompression block uses UMEM as the results buffer for
decompressed weights and one of the SPADs as inputs buffer, and the
inner product would also need two SPADs for inputs and results, to
ensure that the results are not overwritten, we have to send the
partial sums back to the host result buffer after finishing every
strip. The decomp_result_buf is used for that purpose.

Also, we quantize results into half-precision float points before
sending them back to the host results buffer. The problem is that
the pack_and_store_fp16 block sends 16 half-precision elements at
a time (Sam decided to use 16, must have good performance I guess),
meaning that the host results buffer needs to be at least that large.
We didn't take that into account when allocating the fp16 array for
the host result buffer.